requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
argcomplete==3.6.2
asgiref==3.9.1
attr==0.3.1
attrs==25.3.0
azure-core==1.35.0
azure-storage-blob==12.26.0
black==25.1.0
bleach==5.0.1
blinker==1.7.0
boto==2.49.0
boto3==1.39.8
botocore==1.39.8
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.7
colorama==0.4.6
cryptography==45.0.5
datamodel-code-generator==0.26.1
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
Django==5.1.11
django-annoying==0.10.6
django-cors-headers==4.7.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.2.3
django-filter==24.3
django-migration-linter==5.2.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.10.3
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dnspython==2.7.0
dominate==2.9.1
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
email_validator==2.2.0
expiringdict==1.2.2
Faker==37.4.2
filelock==3.18.0
Flask==3.0.0
Flask-Bootstrap==3.3.7.1
fsspec==2025.7.0
genson==1.3.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.33.4
humansignal-drf-yasg==1.21.10.post1
idna==3.10
ijson==3.4.0
importlib_metadata==8.7.0
inflect==5.6.2
inflection==0.5.1
isodate==0.7.2
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.2
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsf==0.11.2
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
label-studio==1.20.0
label-studio-sdk==1.0.18
launchdarkly-server-sdk==8.2.1
lockfile==0.12.2
lxml==6.0.0
lxml_html_clean==0.4.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
mypy_extensions==1.1.0
narwhals==1.47.1
networkx==3.5
nltk==3.9.1
numpy==1.26.4
openai==1.97.0
opencv-python==4.11.0.86
opentelemetry-api==1.35.0
ordered-set==4.0.2
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.7
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyboxen==1.3.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pyRFC3339==2.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==2.0.4
pytz==2022.7.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
requests-mock==1.12.1
rich==14.0.0
rpds-py==0.26.0
rq==1.16.2
rsa==4.9.1
rstr==3.2.2
rules==3.4
s3transfer==0.13.0
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
semver==3.0.4
sentry-sdk==2.33.0
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
sqlparse==0.5.3
streamlit==1.47.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tldextract==5.3.0
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.53.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
ujson==5.10.0
uritemplate==4.2.0
urllib3==1.26.20
user-agents==2.2.0
virtualenv==20.30.0
visitor==0.1.3
watchdog==6.0.0
webencodings==0.5.1
Werkzeug==3.0.1
wrapt==1.17.2
xmljson==0.2.1
zipp==3.23.0
